uint fix

+++ entities/led_74hc595.go
@@ -4,7 +4,7 @@ type Led74HC595 struct {
 	IC *IC_74HC595
 }
 
-func (d *Led74HC595) IsOn(index int) bool {
+func (d *Led74HC595) IsOn(index uint) bool {
 	if index > 7 {
 		return false
 	}
@@ -14,7 +14,7 @@ func (d *Led74HC595) IsOn(index int) bool {
 func (d *Led74HC595) RowStatus() []bool {
 	var r []bool
 	for i := 0; i < 8; i++ {
-		r = append(r, d.IC.Data >> i & 0x01)
+		r = append(r, d.IC.Data >> uint(i) & 0x01)
 	}
 	return r
 }
@@ -27,11 +27,11 @@ func (d *Led74HC595) Off() {
 	d.IC.Clear()
 }
 
-func (d *Led74HC595) OnForIndex(index int) {
+func (d *Led74HC595) OnForIndex(index uint) {
 	d.IC.SetDate(d.IC.Data | (0x01 << index))
 }
 
-func (d *Led74HC595) OffForIndex(index int) {
+func (d *Led74HC595) OffForIndex(index uint) {
 	arr := []uint8{0xfe, 0xfd, 0xfb, 0xf7, 0xef, 0xdf, 0xbf, 0x7f}
 	d.IC.SetDate(d.IC.Data & arr[index])
 }
